move `impl bags_list, im_online, nomination_pools Config for Runtime` to separate file
